+++ regen_openapi.sh
@@ -48,7 +48,7 @@ yarn openapi-generator-cli generate -i openapi.json -g csharp -o csharp/ -c csha
 
 # Cleanup previous codegen, allowing us to spot removals.
 # If the removals are expected, stage them eg. `git add -u`, then commit them.
-rm -rf rust/src/{apis,models}
+rm -rf rust/src/models
 yarn openapi-generator-cli generate -i rust/openapi.json -g rust -o rust/ -c rust/openapi-generator-config.json -t rust/templates
 
 echo Note: Python generation is not executed automatically.